loading-component-updates
feature: Updated Loading component props (iconVariant) to accept a defaulted "primary" variant as well as a "secondary" variant to be provided for different icons to display. The "svg" prop will take priority if provided and will overwrite the variants from displaying. Added tests and small util file along with tests for util.